pyproject.toml
[project]
name = "TRI-LBM"
version = "0.0.7"
description = "Large Behavioral Model from Toyota Research"
authors = [
    { name = "Phil Wang", email = "lucidrains@gmail.com" }
]
readme = "README.md"
requires-python = ">= 3.9"
license = "MIT"
keywords = [
    'artificial intelligence',
    'deep learning',
    'transformers',
    'attention mechanism',
    'diffusion',
    'robotic dexterity'
]

classifiers=[
    'Development Status :: 4 - Beta',
    'Intended Audience :: Developers',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
    'Programming Language :: Python :: 3.9',
]

dependencies = [
    "bidirectional-cross-attention>=0.1.0",
    "denoising-diffusion-pytorch>=2.2.3",
    "einx>=0.3.0",
    "einops>=0.8.0",
    "open_clip_torch",
    "x-transformers>=2.4.9",
    "vit-pytorch>=1.11.2"
]

[project.urls]
Homepage = "https://pypi.org/project/TRI-LBM/"
Repository = "https://github.com/lucidrains/TRI-LBM"

[project.optional-dependencies]
examples = []
test = [
    "pytest"
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
